feat: hostd config simple
