+++ .changeset/chatty-jobs-knock.md
@@ -1,0 +1,5 @@
+---
+'hostd': minor
+---
+
+The configuration now has an advanced mode that allows the user to view and change all settings.
