+++ .changeset/honest-hotels-relate.md
@@ -1,0 +1,5 @@
+---
+'hostd': minor
+---
+
+The configuration page now shows the changed status on fields if the user has made a change but the server values were since updated.
