+++ .changeset/late-nails-type.md
@@ -1,0 +1,5 @@
+---
+'hostd': minor
+---
+
+The command palette now includes navigation to configuration sections.
