+++ .changeset/seven-eagles-pull.md
@@ -1,0 +1,5 @@
+---
+'hostd': minor
+---
+
+The configuration is now much simpler by default, only requiring the user to set essential settings.
