+++ server/Caddyfile-dev
@@ -5,6 +5,7 @@
 # 127.0.0.1 design.local
 # 127.0.0.1 zen.local
 # 127.0.0.1 host.local
+# 127.0.0.1 hostd.local
 # 127.0.0.1 renter.local
 # 127.0.0.1 renterd.local
 # 127.0.0.1 renterd.zen.local
@@ -108,6 +109,11 @@ wallet.local {
 }
 
 renterd.local {
+  import cors
+  reverse_proxy localhost:9980
+}
+
+hostd.local {
   import cors
   reverse_proxy localhost:9980
 }
